requirements.txt
streamlit
python-docx
pdfplumber
pytesseract
pdf2image
requests
python-dotenv
Pillow